[#171487131] Rename file: clients/api.ts -> clients/admin.ts
