+++ src/commands/api-services/create.ts
@@ -7,7 +7,7 @@ import { fromEither, fromPredicate, TaskEither } from "fp-ts/lib/TaskEither";
 // tslint:disable-next-line: no-submodule-imports
 import { getRequiredStringEnv } from "io-functions-commons/dist/src/utils/env";
 import { errorsToReadableMessages } from "italia-ts-commons/lib/reporters";
-import { ApiClient } from "../../clients/api";
+import { ApiClient } from "../../clients/admin";
 import { Service as AdminService } from "../../generated/Service";
 import { errorsToError } from "../../utils/conversions";
